Extracts <SearchContents>
